add associated token account resource

+++ internal/solana/data_source_rent_exemption_cost.go
@@ -32,7 +32,7 @@ func dataSourceRentExemptionCost() *schema.Resource {
 func dataSourceRentExemptionCostRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*providerConfig).rpcClient
 
-	dataLength := d.Get("data_length").(int)
+	dataLength := d.Get("data_length").(uint64)
 	res, err := client.GetMinimumBalanceForRentExemption(context.Background(), dataLength, rpc.CommitmentRecent)
 	if err != nil {
 		return err
